+++ src/main/java/com/prography/lighton/performance/users/application/resolver/PerformanceResolver.java
@@ -13,12 +13,12 @@ import com.prography.lighton.performance.common.domain.entity.vo.Info;
 import com.prography.lighton.performance.common.domain.entity.vo.Location;
 import com.prography.lighton.performance.common.domain.entity.vo.Payment;
 import com.prography.lighton.performance.common.domain.entity.vo.Schedule;
-import com.prography.lighton.performance.users.presentation.dto.InfoDTO;
-import com.prography.lighton.performance.users.presentation.dto.PaymentDTO;
-import com.prography.lighton.performance.users.presentation.dto.RegisterPerformanceMultiPart;
-import com.prography.lighton.performance.users.presentation.dto.SavePerformanceRequest;
-import com.prography.lighton.performance.users.presentation.dto.ScheduleDTO;
-import com.prography.lighton.performance.users.presentation.dto.UpdatePerformanceMultiPart;
+import com.prography.lighton.performance.users.presentation.dto.request.InfoDTO;
+import com.prography.lighton.performance.users.presentation.dto.request.PaymentDTO;
+import com.prography.lighton.performance.users.presentation.dto.request.RegisterPerformanceMultiPart;
+import com.prography.lighton.performance.users.presentation.dto.request.SavePerformanceRequest;
+import com.prography.lighton.performance.users.presentation.dto.request.ScheduleDTO;
+import com.prography.lighton.performance.users.presentation.dto.request.UpdatePerformanceMultiPart;
 import com.prography.lighton.region.application.dto.Coordinate;
 import com.prography.lighton.region.application.service.AddressGeocodingService;
 import com.prography.lighton.region.infrastructure.cache.RegionCache;
