+++ src/main/java/com/prography/lighton/performance/users/application/service/ArtistPerformanceService.java
@@ -7,8 +7,8 @@ import com.prography.lighton.performance.common.domain.entity.Performance;
 import com.prography.lighton.performance.common.domain.entity.enums.PerformanceFilterType;
 import com.prography.lighton.performance.users.application.resolver.PerformanceResolver;
 import com.prography.lighton.performance.users.infrastructure.repository.PerformanceRepository;
-import com.prography.lighton.performance.users.presentation.dto.RegisterPerformanceMultiPart;
-import com.prography.lighton.performance.users.presentation.dto.UpdatePerformanceMultiPart;
+import com.prography.lighton.performance.users.presentation.dto.request.RegisterPerformanceMultiPart;
+import com.prography.lighton.performance.users.presentation.dto.request.UpdatePerformanceMultiPart;
 import com.prography.lighton.performance.users.presentation.dto.response.GetPerformanceMapListResponseDTO;
 import com.prography.lighton.performance.users.presentation.dto.response.PerformanceSearchItemDTO;
 import java.time.LocalDate;
