+++ src/main/java/com/prography/lighton/performance/users/application/service/BuskingService.java
@@ -6,10 +6,10 @@ import com.prography.lighton.member.common.domain.entity.Member;
 import com.prography.lighton.performance.common.domain.entity.Busking;
 import com.prography.lighton.performance.users.application.resolver.PerformanceResolver;
 import com.prography.lighton.performance.users.infrastructure.repository.PerformanceRepository;
-import com.prography.lighton.performance.users.presentation.dto.ArtistBuskingRegisterRequest;
-import com.prography.lighton.performance.users.presentation.dto.ArtistBuskingUpdateRequest;
-import com.prography.lighton.performance.users.presentation.dto.UserBuskingRegisterRequest;
-import com.prography.lighton.performance.users.presentation.dto.UserBuskingUpdateRequest;
+import com.prography.lighton.performance.users.presentation.dto.request.ArtistBuskingRegisterRequest;
+import com.prography.lighton.performance.users.presentation.dto.request.ArtistBuskingUpdateRequest;
+import com.prography.lighton.performance.users.presentation.dto.request.UserBuskingRegisterRequest;
+import com.prography.lighton.performance.users.presentation.dto.request.UserBuskingUpdateRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
