+++ src/main/java/com/prography/lighton/performance/users/presentation/controller/ArtistBuskingController.java
@@ -5,8 +5,8 @@ import com.prography.lighton.common.utils.ApiUtils;
 import com.prography.lighton.common.utils.ApiUtils.ApiResult;
 import com.prography.lighton.member.common.domain.entity.Member;
 import com.prography.lighton.performance.users.application.service.BuskingService;
-import com.prography.lighton.performance.users.presentation.dto.ArtistBuskingRegisterRequest;
-import com.prography.lighton.performance.users.presentation.dto.ArtistBuskingUpdateRequest;
+import com.prography.lighton.performance.users.presentation.dto.request.ArtistBuskingRegisterRequest;
+import com.prography.lighton.performance.users.presentation.dto.request.ArtistBuskingUpdateRequest;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
