+++ src/main/java/com/prography/lighton/performance/users/presentation/controller/PerformanceController.java
@@ -5,9 +5,9 @@ import com.prography.lighton.common.utils.ApiUtils;
 import com.prography.lighton.common.utils.ApiUtils.ApiResult;
 import com.prography.lighton.member.common.domain.entity.Member;
 import com.prography.lighton.performance.users.application.service.ArtistPerformanceService;
-import com.prography.lighton.performance.users.presentation.dto.RegisterPerformanceMultiPart;
-import com.prography.lighton.performance.users.presentation.dto.SavePerformanceRequest;
-import com.prography.lighton.performance.users.presentation.dto.UpdatePerformanceMultiPart;
+import com.prography.lighton.performance.users.presentation.dto.request.RegisterPerformanceMultiPart;
+import com.prography.lighton.performance.users.presentation.dto.request.SavePerformanceRequest;
+import com.prography.lighton.performance.users.presentation.dto.request.UpdatePerformanceMultiPart;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotNull;
 import lombok.RequiredArgsConstructor;
