+++ src/main/java/com/prography/lighton/performance/users/presentation/controller/UserBuskingController.java
@@ -5,8 +5,8 @@ import com.prography.lighton.common.utils.ApiUtils;
 import com.prography.lighton.common.utils.ApiUtils.ApiResult;
 import com.prography.lighton.member.common.domain.entity.Member;
 import com.prography.lighton.performance.users.application.service.BuskingService;
-import com.prography.lighton.performance.users.presentation.dto.UserBuskingRegisterRequest;
-import com.prography.lighton.performance.users.presentation.dto.UserBuskingUpdateRequest;
+import com.prography.lighton.performance.users.presentation.dto.request.UserBuskingRegisterRequest;
+import com.prography.lighton.performance.users.presentation.dto.request.UserBuskingUpdateRequest;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
