+++ src/main/java/com/prography/lighton/performance/users/presentation/dto/request/RegisterPerformanceMultiPart.java
@@ -1,9 +1,9 @@
-package com.prography.lighton.performance.users.presentation.dto;
+package com.prography.lighton.performance.users.presentation.dto.request;
 
 import jakarta.validation.Valid;
 import org.springframework.web.multipart.MultipartFile;
 
-public record UpdatePerformanceMultiPart(
+public record RegisterPerformanceMultiPart(
         @Valid SavePerformanceRequest data,
         MultipartFile posterImage,
         MultipartFile proof
