+++ src/main/java/com/prography/lighton/performance/users/presentation/dto/request/UserBuskingRegisterRequest.java
@@ -1,11 +1,11 @@
-package com.prography.lighton.performance.users.presentation.dto;
+package com.prography.lighton.performance.users.presentation.dto.request;
 
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import org.hibernate.validator.constraints.URL;
 
-public record UserBuskingUpdateRequest(
+public record UserBuskingRegisterRequest(
 
         @NotNull(message = "공연 기본 정보는 필수입니다.")
         @Valid
